printf %x

+++ lib/include/uart.h
@@ -1,6 +1,8 @@
 #ifndef _UART_H
 #define _UART_H
 
+#include "int.h"
+
 void uart_init();
 void uart_write(unsigned int c);
 
@@ -12,6 +14,6 @@ void uart_print(char *s);
 void uart_println(char *s);
 
 /* Display a binary value in hexadecimal */
-void uart_hex(unsigned int d);
+void uart_hex(size_t d);
 
-#endif // _UART_H
+#endif  // _UART_H
